+++ src/main/java/io/github/system_false/random/GenerationException.java
@@ -15,7 +15,7 @@
  *
  */
 
-package org.system_false.random.generator;
+package io.github.system_false.random;
 
 /**
  * Thrown when there is an error while generating random value.
