+++ src/main/java/io/github/system_false/random/RandomValue.java
@@ -15,7 +15,7 @@
  *
  */
 
-package org.system_false.random.generator;
+package io.github.system_false.random;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -32,7 +32,6 @@ import java.lang.annotation.Target;
  *     <li>{@code <type>MaxValue}: maximum value of primitive type. Example: {@code intMaxValue = 10}.</li>
  *     <li>{@code <type>Values}: possible values of primitive type. Example: {@code intValues = {0, 1, 2}}.</li>
  * </ul>
- * </p>
  * <p>
  * The following parameters are used to configure randomization of string:
  * <ul>
@@ -43,28 +42,24 @@ import java.lang.annotation.Target;
  *     <li>{@code charMaxValue}: maximum char value in string</li>
  *     <li>{@code charValues}: possible characters in string. Example: {@code charValues = {'a', 'b', 'c'}}</li>
  * </ul>
- * </p>
  * <p>
  * The following parameters are used to configure randomization of enums:
  * <ul>
  *     <li>{@code defaultEnum}: default value of enum. Example: {@code defaultEnum = "A"}</li>
  *     <li>{@code enumValues}: possible names of enum. Example: {@code enumValues = {"A", "B", "C"}}</li>
  * </ul>
- * </p>
  * <p>
  * The following parameters are used to configure randomization of arrays:
  * <ul>
  *     <li>{@code arrayMinSize}: minimum size of array</li>
  *     <li>{@code arrayMaxSize}: maximum size of array</li>
  * </ul>
- * </p>
  * <p>
  * The following parameters are used to configure randomization of lists, sets and maps:
  * <ul>
  *     <li>{@code containerMinSize}: minimum size of container</li>
  *     <li>{@code containerMaxSize}: maximum size of container</li>
  * </ul>
- * </p>
  * <p>
  * Parameter {@code objectDepth} describes the depth of the object to be generated. Simply, it specifies how many
  * nested objects will be generated by random. The default value is 3.
@@ -235,14 +230,14 @@ public @interface RandomValue {
 
     /**
      * Minimal value of float, must be less than or equal to {@link #floatMaxValue()}. This
-     * value is ignored if {@link #floatValues()} is specified. Default is {@link 0}.
+     * value is ignored if {@link #floatValues()} is specified. Default is {@code 0}.
      * @return the minimal float value
      */
     float floatMinValue() default 0F;
 
     /**
      * Maximal value of float, must be greater than or equal to {@link #floatMinValue()}. This
-     * value is ignored if {@link #floatValues()} is specified. Default is {@link 1}.
+     * value is ignored if {@link #floatValues()} is specified. Default is {@code 1}.
      * @return the maximal float value
      */
     float floatMaxValue() default 1F;
@@ -262,14 +257,14 @@ public @interface RandomValue {
 
     /**
      * Minimal value of double, must be less than or equal to {@link #doubleMaxValue()}. This
-     * value is ignored if {@link #doubleValues()} is specified. Default is {@link 0}.
+     * value is ignored if {@link #doubleValues()} is specified. Default is {@code 0}.
      * @return the minimal double value
      */
     double doubleMinValue() default 0F;
 
     /**
      * Maximal value of double, must be greater than or equal to {@link #doubleMinValue()}. This
-     * value is ignored if {@link #doubleValues()} is specified. Default is {@link 1}.
+     * value is ignored if {@link #doubleValues()} is specified. Default is {@code 1}.
      * @return the maximal double value
      */
     double doubleMaxValue() default 1F;
@@ -289,14 +284,14 @@ public @interface RandomValue {
 
     /**
      * Minimal length of string, must be less than or equal to {@link #stringMaxLength()}. This
-     * value is ignored if {@link #stringValues()} is specified. Default is {@link 1}.
+     * value is ignored if {@link #stringValues()} is specified. Default is {@code 1}.
      * @return the minimal string length
      */
     int stringMinLength() default 1;
 
     /**
      * Maximal length of string, must be greater than or equal to {@link #stringMinLength()}. This
-     * value is ignored if {@link #stringValues()} is specified. Default is {@link 1}.
+     * value is ignored if {@link #stringValues()} is specified. Default is {@code 1}.
      * @return the maximal string length
      */
     int stringMaxLength() default 1;
