+++ src/main/java/io/github/system_false/random/RangeGenerator.java
@@ -15,7 +15,7 @@
  *
  */
 
-package org.system_false.random.generator;
+package io.github.system_false.random;
 
 import java.util.function.Predicate;
