+++ src/main/java/module-info.java
@@ -1,3 +1,9 @@
+/**
+ * Module of Random Library. It contains classes for generating random objects.
+ * It can be used to generate random values for primitive types, String, enum, record,
+ * array, List, Set, Map and custom classes with {@link io.github.system_false.random.RandomValue}
+ * annotation and {@link io.github.system_false.random.RandomCreator} annotations.
+ */
 module random_lib {
-    exports org.system_false.random.generator;
+    exports io.github.system_false.random;
 }
